// src/types/common.ts
export type TicketData = {
  origin: string;
  origin_name: string;
  destination: string;
  destination_name: string;
  departure_date: string;
  departure_time: string;
  arrival_date: string;
  arrival_time: string;
  carrier: string;
  stops: number;
  price: number;
  id: string;
};

export interface Stops {
  [key: string]: boolean;
}

export type Currency = 'Rub' | 'Usd' | 'Eur';